# OBPIH-7368 Improve performance of export/import of counting step with…