--- a/src/js/api/services/CycleCountApi.js
+++ b/src/js/api/services/CycleCountApi.js
@@ -3,6 +3,7 @@ import queryString from 'query-string';
 import {
   CYCLE_COUNT, CYCLE_COUNT_ITEM,
   CYCLE_COUNT_ITEMS_BATCH,
+  CYCLE_COUNT_ITEMS_BATCH_ROOT,
   CYCLE_COUNT_ITEMS_IMPORT,
   CYCLE_COUNT_PENDING_REQUESTS,
   CYCLE_COUNT_RECOUNT_START,
@@ -52,8 +53,14 @@ export default {
       removeOutOfStockItemsImplicitly), {}, { params: { countIndex } }),
   createCycleCountItems: (payload, locationId, cycleCountId) =>
     apiClient.post(CYCLE_COUNT_ITEMS_BATCH(locationId, cycleCountId), payload),
+  // Root endpoint doesn't require the cycleCountId - we can send items from multiple cycle counts
+  createCycleCountItemsBatch: (payload, locationId) =>
+    apiClient.post(CYCLE_COUNT_ITEMS_BATCH_ROOT(locationId), payload),
   updateCycleCountItems: (payload, locationId, cycleCountId) =>
     apiClient.patch(CYCLE_COUNT_ITEMS_BATCH(locationId, cycleCountId), payload),
+  // Root endpoint doesn't require the cycleCountId - we can send items from multiple cycle counts
+  updateCycleCountItemsBatch: (payload, locationId) =>
+    apiClient.patch(CYCLE_COUNT_ITEMS_BATCH_ROOT(locationId), payload),
   importCycleCountItems: (file, locationId) => {
     const formData = new FormData();
     formData.append('importFile', file);