--- a/src/main/groovy/org/pih/warehouse/inventory/CycleCountItemCommand.groovy
+++ b/src/main/groovy/org/pih/warehouse/inventory/CycleCountItemCommand.groovy
@@ -54,7 +54,7 @@ class CycleCountItemCommand implements Validateable {
     def beforeValidate() {
         String cycleCountId = RequestContextHolder.getRequestAttributes().params?.cycleCountId
         String facilityId = RequestContextHolder.getRequestAttributes().params?.facility
-        cycleCount = CycleCount.get(cycleCountId)
+        cycleCount = cycleCount ?: CycleCount.get(cycleCountId)
         facility = Location.read(facilityId)
     }